gcs: Separate http methods into their own file
